force color_vector to be pd.Categorical
The `utils.py` function `_set_color_source_vec` generates `color_source_vector`
and `color_vector` -- the former contains the label of each element, the latter
the color. This is done by first extracting a `color_source_vector`
`pd.Series` of `Categorical` dtype, then defining a `color_mapping` and then
calling `color_source_vector.map(color_mapping)` to generate the
`color_vector`. `pd.Series.map` creates a Categorical Series if
`color_mapping` is unique, but a pd.Index if `color_mapping` is not unique.
However, we later (in `_render_labels`) check if `color_vector` is Categorical,
and error out if it isn't. Therefore, we should prevent it from being returned
as pd.Index in the first place.